+++ routes/api.php
@@ -4,6 +4,10 @@ use Illuminate\Support\Facades\Route;
 
 Route::middleware('auth:sanctum')->prefix('v1')->group(function(){
 
+    Route::prefix('user')->group(function(){
+        require __DIR__ . '/api/user.php';
+    });
+
 });
 
 Route::prefix('v1')->group(function(){